Custom Exception

+++ Screamer.Application/Features/PostRequest/Commands/DeletePostRequest/DeletePostRequestHandlerCommand.cs
@@ -4,7 +4,9 @@ using System.Linq;
 using System.Threading.Tasks;
 using AutoMapper;
 using MediatR;
+using Screamer.Application.Contracts.Exceptions;
 using Screamer.Application.Contracts.Presistance;
+using Screamer.Domain.Common;
 
 namespace Screamer.Application.Features.PostRequest.Commands.DeletePostRequest
 {
@@ -28,7 +30,10 @@ namespace Screamer.Application.Features.PostRequest.Commands.DeletePostRequest
       var PostToDelete = await _postRepository.GetByIdAsync(request.Id);
 
             // verify that record exists
-        
+            if (PostToDelete == null)
+            {
+                throw new NotFoundException(nameof(Post), request.Id);
+            }
 
             // remove from database
             await _postRepository.DeleteAsync(PostToDelete);
